deploy.py
#!/usr/bin/env python3
"""
Auto-deploy script for LegalRAG
Creates all project files in the correct structure
Run: python deploy.py
"""

import os
import json
from pathlib import Path

# ==============================================================================
# FILE CONTENTS DICTIONARY
# ==============================================================================

FILES = {
    # ==================== CONFIGURATION FILES ====================
    "requirements.txt": """torch==2.0.1
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
numpy==1.24.3
pytest==7.4.3
pytest-cov==4.1.0""",

    "pyproject.toml": """[project]
name = "legalrag"
version = "0.1.0"
description = "RAG-based legal document assistant"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100""",

    ".gitignore": """# Python
__pycache__/
*.py[cod]
*$py.class
*.so
.Python
build/
develop-eggs/
dist/
downloads/
eggs/
.eggs/
lib/
lib64/
parts/
sdist/
var/
wheels/
*.egg-info/
.installed.cfg
*.egg

# Virtual environments
venv/
ENV/
env/

# IDE
.vscode/
.idea/
*.swp
*.swo
*~
.DS_Store

# Models and large files
*.bin
*.model
*.pt
*.pth
*.ckpt
data/raw/*.json
data/processed/*.jsonl
data/indices/*.bin
*.pkl

# Jupyter
.ipynb_checkpoints/
*.ipynb_checkpoints

# Testing
.pytest_cache/
.coverage
htmlcov/

# Logs
*.log
logs/""",

    # ==================== MAIN SOURCE FILES ====================
    "src/__init__.py": '"""LegalRAG: RAG-based legal document assistant."""\n__version__ = "0.1.0"',

    "src/config.py": """\"\"\"Configuration module for LegalRAG system.\"\"\"
import os
from dataclasses import dataclass
from pathlib import Path
from typing import Optional
import yaml

PROJECT_ROOT = Path(__file__).parent.parent
DATA_DIR = PROJECT_ROOT / "data"
MODELS_DIR = PROJECT_ROOT / "models"
CONFIGS_DIR = PROJECT_ROOT / "configs"

@dataclass
class ModelConfig:
    \"\"\"Configuration for LLM and embedding models.\"\"\"
    embedding_model_name: str = "sentence-transformers/all-MiniLM-L6-v2"
    embedding_dim: int = 384
    
    llm_model_name: str = "mistralai/Mistral-7B-Instruct-v0.2"
    llm_max_tokens: int = 512
    llm_temperature: float = 0.3
    llm_top_p: float = 0.9
    
    device: str = "cpu"
    quantization: bool = False

@dataclass
class RAGConfig:
    \"\"\"Configuration for RAG pipeline.\"\"\"
    chunk_size: int = 512
    chunk_overlap: int = 100
    top_k: int = 3
    similarity_threshold: float = 0.5
    
    index_type: str = "faiss"
    metric_type: str = "l2"
    
    max_source_tokens: int = 2000
    system_prompt_template: str = "legal"
    
    enable_safety_checks: bool = True
    check_hallucination: bool = True
    max_refusal_rate: float = 0.1

@dataclass
class DataConfig:
    \"\"\"Configuration for data handling.\"\"\"
    raw_data_path: Path = DATA_DIR / "raw" / "sample_legal_docs.json"
    processed_data_path: Path = DATA_DIR / "processed" / "chunks.jsonl"
    index_path: Path = DATA_DIR / "indices" / "faiss_index.bin"
    metadata_path: Path = DATA_DIR / "indices" / "metadata.json"
    
    test_split: float = 0.1
    val_split: float = 0.1
    random_seed: int = 42

@dataclass
class APIConfig:
    \"\"\"Configuration for FastAPI server.\"\"\"
    host: str = "0.0.0.0"
    port: int = 8000
    reload: bool = True
    workers: int = 1
    timeout: float = 30.0

@dataclass
class EvaluationConfig:
    \"\"\"Configuration for evaluation.\"\"\"
    num_eval_samples: int = 20
    metrics: list = None
    
    def __post_init__(self):
        if self.metrics is None:
            self.metrics = ["retrieval_recall@3", "retrieval_precision@3", "generation_quality"]

class AppConfig:
    \"\"\"Main application configuration.\"\"\"
    
    def __init__(self, config_path: Optional[str] = None):
        self.model = ModelConfig()
        self.rag = RAGConfig()
        self.data = DataConfig()
        self.api = APIConfig()
        self.evaluation = EvaluationConfig()
        
        if config_path:
            self.load_from_yaml(config_path)
    
    def load_from_yaml(self, config_path: str):
        \"\"\"Load configuration from YAML file.\"\"\"
        with open(config_path, 'r') as f:
            config_dict = yaml.safe_load(f)
        
        if 'model' in config_dict:
            self.__dict__['model'] = ModelConfig(**config_dict['model'])
        if 'rag' in config_dict:
            self.__dict__['rag'] = RAGConfig(**config_dict['rag'])
        if 'data' in config_dict:
            self.__dict__['data'] = DataConfig(**config_dict['data'])
        if 'api' in config_dict:
            self.__dict__['api'] = APIConfig(**config_dict['api'])
        if 'evaluation' in config_dict:
            self.__dict__['evaluation'] = EvaluationConfig(**config_dict['evaluation'])

def get_config(config_path: Optional[str] = None) -> AppConfig:
    \"\"\"Get application configuration.\"\"\"
    return AppConfig(config_path)
""",

    # [Здесь остальные файлы из раздела C - очень много текста]
    # Для экономии места в этом скрипте покажу структуру с основными файлами
    
}

# ==============================================================================
# DEPLOYMENT LOGIC
# ==============================================================================

def create_project_structure():
    """Create all project directories."""
    dirs = [
        "src",
        "api",
        "data/raw",
        "data/processed",
        "data/indices",
        "configs",
        "notebooks",
        "tests",
        "scripts",
        "reports",
        "docs",
        "presentation"
    ]
    
    for d in dirs:
        Path(d).mkdir(parents=True, exist_ok=True)
        print(f"✓ Created directory: {d}")

def create_init_files():
    """Create __init__.py files."""
    init_files = ["src/__init__.py", "api/__init__.py", "tests/__init__.py"]
    
    for f in init_files:
        Path(f).touch()
        print(f"✓ Created: {f}")

def write_files():
    """Write all configuration files."""
    for filepath, content in FILES.items():
        Path(filepath).parent.mkdir(parents=True, exist_ok=True)
        with open(filepath, 'w', encoding='utf-8') as f:
            f.write(content)
        print(f"✓ Created: {filepath}")

def main():
    print("=" * 70)
    print("🚀 LegalRAG Auto-Deploy Script")
    print("=" * 70)
    print()
    
    print("📁 Creating project structure...")
    create_project_structure()
    print()
    
    print("📝 Creating configuration files...")
    write_files()
    print()
    
    print("=" * 70)
    print("✅ Project initialized successfully!")
    print("=" * 70)
    print()
    print("📋 Next steps:")
    print("1. Copy the source files from the report into their directories")
    print("2. pip install -r requirements.txt")
    print("3. python scripts/generate_synthetic_data.py")
    print("4. python scripts/build_index.py")
    print("5. pytest tests/ -v")
    print("6. python scripts/run_server.py")
    print()
    print("🌐 Access API at: http://localhost:8000/docs")
    print()

if __name__ == "__main__":
    main()
